Adds prototype validation for transcripts
Adds a backend check using a python script that had been provided by Andrew.
This is not safe, it should not be released, it's simply a POC that the script does work.
If this will be used in production then it needs to be refactored out so user input
is not being passed to exec. Preferably exec would not be called at all.